Move remaining parameter parser functions
